% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertices_utils.R
\name{add_vertices}
\alias{add_vertices}
\title{add_vertices}
\usage{
add_vertices(sf, each = 0.1, parallel = FALSE, ...)
}
\arguments{
\item{sf}{an object of class \code{sf}}

\item{each}{the step value to use to create vertices}

\item{parallel}{run in parallel}

\item{...}{parallel options}
}
\value{
an object of class \code{sf}
}
\description{
Densify a spatial polygon by adding vertices
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
