% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertices_utils.R
\name{bearing_between_vertices}
\alias{bearing_between_vertices}
\title{bearing_between_vertices}
\usage{
bearing_between_vertices(features)
}
\arguments{
\item{sf}{an object of class \code{sf}}
}
\value{
an object of class \code{data.frame} including vertices indexes and the bearing
}
\description{
Calculates the bearing between each pair of coordinate. The bearing
is computed with the \link[geosphere]{bearingRhumb} function.
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
