% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_utils.R
\name{create_cwp_grid}
\alias{create_cwp_grid}
\title{create_cwp_grid}
\usage{
create_cwp_grid(
  size = NULL,
  res = NULL,
  xmin = NULL,
  ymin = NULL,
  xmax = NULL,
  ymax = NULL,
  densify = FALSE,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{size}{an integer code corresponding to the grid size (referred as code
A in the CWP Handbook)}

\item{res}{a string matching one of the accepted resolution values. Accepted 
resolutions values are '10min_x_10min', '20min_x_20min', '30min_x_30min',
'30min_x_1deg', '1deg_x_1deg', '5deg_x_5deg', '10deg_x_10deg', '20deg_x_20deg',
'30deg_x_30deg'"}

\item{xmin}{xmin of the output grid}

\item{ymin}{ymin of the output grid}

\item{xmax}{xmax of the output grid}

\item{ymax}{ymax of the output grid}

\item{densify}{densify}

\item{parallel}{run in parallel}

\item{...}{parallel options}
}
\value{
an object of class \code{sf}
}
\description{
Creates a CWP grid spatial object
}
\references{
CWP Handbook - https://www.fao.org/cwp-on-fishery-statistics/handbook/general-concepts/main-water-areas/fr/#c737133
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
