% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_geodesic_buffer.R
\name{create_geodesic_buffer}
\alias{create_geodesic_buffer}
\title{create_geodesic_buffer}
\usage{
create_geodesic_buffer(
  features,
  distance,
  unit = c("m", "nm"),
  n_segments = 72
)
}
\arguments{
\item{features}{an object of class \code{sf}}

\item{distance}{distance}

\item{unit}{unit for the distance (either 'm' for meters, or 'nm' for nautical miles)}

\item{n_segments}{number of segments for each single vertex buffer (intermediate step).
By default 72 which means a segment for each 5 degree bearing. For fine-grained buffers, values
could be 180 (each 2 degree) or 360 (each degree).}
}
\value{
an object of class \code{sf}
}
\description{
Creates a geodesic buffer
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
