% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_loxodrome_line.R
\name{create_loxodrome_line}
\alias{create_loxodrome_line}
\title{create_loxodrome_line}
\usage{
create_loxodrome_line(features, distance = 1852)
}
\arguments{
\item{features}{an lines object of class \code{sf}}

\item{distance}{distance in meters. Default is \code{1852} (1 nautical mile)}
}
\value{
an object of class \code{sf}
}
\description{
Creates a loxodrome line from a straight line
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
