% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertices_utils.R
\name{distance_between_vertices}
\alias{distance_between_vertices}
\title{distance_between_vertices}
\usage{
distance_between_vertices(features)
}
\arguments{
\item{sf}{an object of class \code{sf}}
}
\value{
an object of class \code{data.frame} including vertices indexes and the distance
}
\description{
Calculates the distance between each pair of coordinate. The distance
is computed with the \link[geosphere]{distGeo} function.
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
