% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdisf.R
\docType{package}
\name{fdisf}
\alias{fdisf-package}
\alias{fdisf}
\title{Spatial features for fisheries data interoperability}
\description{
fdisf provides a set of spatial features for fisheries data interoperability
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fdiwg/fdisf}
  \item Report bugs at \url{https://github.com/fdiwg/fdisf/issues}
}

}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
