% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_utils.R
\docType{data}
\name{grid_char_5x5}
\alias{grid_char_5x5}
\title{A constant representing the initial character of a 5x5 regular grid code}
\format{
An object of class \code{character} of length 1.
}
\usage{
grid_char_5x5
}
\description{
A constant representing the initial character of a 5x5 regular grid code
}
\keyword{datasets}
