% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_utils.R
\name{optimize_bbox}
\alias{optimize_bbox}
\title{optimize_bbox}
\usage{
optimize_bbox(sf)
}
\arguments{
\item{sf}{object of class \pkg{sf}}
}
\value{
an object of class \code{bbox}
}
\description{
Creates an optimized bbox for an \pkg{sf} object. The optimization
consists in extending the bounding box to manage features crossing the dateline.
}
\author{
Emmanuel Blondel \email{emmanuel.blondel1@gmail.com}
}
